requirements.txt
numpy
streamlit
Pillow
math
opencv-python